requirements.txt
# reqs added for AK's work
fastapi
pydantic
uvicorn
python-multipart

# deps added for BI's work
chromadb==1.3.7
pypdf
groq

# deps added for RA's work


# deps added for ST's work
python-jose[cryptography]
# passlib[bcrypt] removed
bcrypt
python-dotenv
pandas
openpyxl
tabulate
sqlalchemy
google-generativeai

